refactor(bootx64): replace with `x86_64` (#119)
Previously, the calling convention of the external assembly functions
was `C`. The mismatch of the calling conventions caused a corrupt binary
for targets that are not System V AMD64 targets, like Windows-like ones,
because the arguments are not passed correctly. Since the recent pull
request fixed this problem, let's replace the manual implementations
with the library's code.