refactor(uefi_wrapper): remove an unused member
When I added the parameter, I did not understand that Rust's borrowing rule prohibited creating multiple mutable references of this type from the `con_out` method without the unused parameter. However, the actual is that the parameter is not necessary. Even without it, the borrowing rule prevents it, so the parameter is not needed.