chore(process): wrap the kernel stack with `Kbop` (#223)
This is necessary for the kernel privilege processes (except the idle process). These processes use the same stack for the process stack and the kernel stack because an interrupt does not change the stack pointer to the kernel one as they are kernel privileges. Using an array as the kernel stack is impossible because it is impossible to get the address of the field of a struct before creating the struct. Therefore we must wrap the kernel stack with `Kbox`.