fix: avoid deadlocks (#259)
* fix: avoid the deadlock on IPC

This commit removes the cause of the deadlock caused by IPC. When a
kernel-privileged process calls `send` or `receive`, the interrupts are
still enabled. If the timer interrupt happens during the executions of
these functions, a process switch occurs with the process manager
locked. Then, if another process (either kernel or user privileged)
calls an IPC function, it fails to lock the process manager because
the preceding kernel-privileged process has already locked it. This
commit changes `send` and `receive` so that these functions disable
interrupts before the actual sending or receiving, then reenable
after the executions. It prevents the deadlocks because no process
switch happens during `call` or `receive.

* docs: add notes why we disable interrupts

* fix: avoid deadlocks caused by `qemu_println`

See the comment writte in `io.rs`.

* style: make clippy happy

* fix: remove an unnecessary log