fix: `receive` did not handle sender pid queue
This commit fixes the problem of not modifying `sending_to_this`
correctly. The `receive` method popped an element from the queue when
the process specified `Any` for the sender but did nothing when the
process determined the sender's PID. This commit rewrote the code so
that `receive` always tries to pop a PID from the queue.